pyproject.toml
[project]
name = "ollaagent"
version = "0.1.0"
description = "Local LLM agent powered by ollama — memory, plan mode, subagents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ollama==0.6.1",
    "pydantic==2.12.5",
    "python-dotenv==1.2.2",
    "pyyaml>=6.0",
    "rich==14.3.3",
]

[project.scripts]
ollaagent = "ollaAgent.agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ollaAgent"]

[dependency-groups]
dev = [
    "black>=24.0.0",
    "isort>=5.13.0",
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
